Character Animations - Section 3
Created a custom Anim Instance class in c++, then created an Animation Blueprint based on that class. Created blend spaces for animations including running and strafing. Created a HUD class to draw crosshairs on the screen during gameplay. Added impact and beam particles to make the shooting of the weapon more realistic. Added a line trace from the tip of the gun barrel so player is now able to aim the character to shoot at a specific target.