Aiming and Spreading Crosshairs - Section 4
Added aiming function to the action mappings resulting in a zoom effect when aiming button is pressed. Setup interpolation for the zoom for a proper transition from normal view to zoomed in view. Updated the animation blueprint to change to an aiming pose when the aim button is pressed. Changed the crosshairs to spread based on the character moving, aiming, shooting and jumping. Crosshairs spread either individually or a summary of all actions that happen simultaneously. Finally I coded a timer to keep track of  bullet fire to simulate automatic gun fire.