The Weapon Class - Section 5
Created a parent Item class and made the Weapon class a child of the Item class. For now the mesh used in this Weapon class is the Submachine Gun (SMG) that came Lt. Belica who is the default character in our game. Created a PickupWidget to be displayed with the Weapon is traced under the crosshairs. Added an action binding that allows the Weapon to be swapped with other Weapons when the 'E' key is pressed (during line trace). In the widget, created function binds to certain aspects of the widget including ItemName, AmmoCount, and Star count. Created an enum (short for enumeration) for the different states of the Weapon to govern behavior during gameplay. Finally created functions in the Character class to handle swapping the equipping the weapon when the 'E' key is pressed.