Item Interpolation - Section 6
Created a couple of curve floats on the Weapon blueprint that allows weapons to interpolate across the screen right above the character when weapon's are being equipped. Also added properties for the EquipInterping state. Other items can also use this code to interpolate (for example ammo will do something similar when being picked up.  Not much coding needed at all for this task.