Weapon Reloading - Section 7
Retargeted the reload animation from the Animation Startup Pack and targeted it to the Belica Character. Added notifications to move the clip of the Weapon during reloading. Also added notifications to play sounds with the animation for a better experience. Created a Player Controller that takes care of the HUD to display the amount of ammo during firing and reload. Also added sounds for equipping and swapping weapons found in the level.