Weapon Glow Effects - Section 10
Created a Post Process material to create a color outline on the Weapon and Ammo blueprints. Added custom depth and glow materials so the Items flash when sitting on the ground (EItemState::Pickup) and when interpolation when the Item is picked up (EItemState::EquipInterping). Blended several materials together combined with Fresnel Effect to have a nice color glow on the Weapons.