Weapon Inventory Bar - Section 10
Added a an Inventory Bar widget to the HUD using F key for default weapon and numbers one through five for those respective spots. Using a delegate in C++ to bind the Inventory array, I created an animation that highlights available spaces on the inventory bar when hovering over a weapon. Also used another delegate to play a small animation that elevates the currently selected weapon in Inventory.