Multiple Weapon Types - Section 11
Created data table for the different weapon types to load all functions of each weapon when equipped.  The three weapon types are Submachine Gun, Assault Rifle and Pistol. Added a FABRIK IK to each weapon type for proper holding of the character. Each weapon also has it's own unique crosshairs also added to the data table. Animation blueprints were completed for Assault Rifle (to handle reloading) and for the Pistol (to handle slide action when firing). Glow material now works on all three weapon types. Added a variable for semi automatic fire which is true for the SMG and AR and false for the Pistol.