AI and Behavior Trees - Section 15
Started by adding the AI basics including AI Controller, AI Blackboard and AI Behavior Tree. Coded several behaviors for the EnemyBP including patrolling, attacking and wait times. Created an Enemy Animation blueprint to handle actions for logic from the AI Controller. Added death animations to both the Enemy and the Character blueprints. Added weapon trails to the Grux enemy template and then added several versions of the Grux enemy with different sizes, health and damage. This commit is WAY oversimplified as a LOT of coding was done to complete the Grux enemy class.