Khaimera Enemy Skins - Section 16
Completed new montages (HitReact, Attack and Death) for the Khaimera character and added all of the available skins that come with him. Also added the ability for the explosives to cause damage to nearby Actors (including Character and Enemies). Also added health pickups to boost Character's health when it gets to low but will never boost past the MaxHealth. This completes all the elements in the game required to start building the final play level! Yay!